Move GristType.SpawnCategory out of GristType
